move substreams docs to own sidebar section
