+++ website/pages/[locale]/index.mdx
@@ -7,8 +7,8 @@ import { buildDynamicMDX } from 'nextra/remote'
 import { RemoteContent } from 'nextra/data'
 import { supportedLocales, translations } from '@/i18n'
 
-import { Intro, NetworkRoles, Products, SupportedNetworks } from '@/components/IndexPage'
-import { getPageMap } from '@/components/get-page-map'
+import { Intro, NetworkRoles, Products, SupportedNetworks } from '@/src/IndexPage'
+import { getPageMap } from '@/src/get-page-map'
 
 export const getStaticPaths = () => {
   return {
