+++ website/pages/_app.mdx
@@ -2,4 +2,4 @@
 
 {/* _app.mdx is used in nextra for better performance and smallest bundle size. */}
 
-export { default } from '../components/_app'
+export { default } from '../src/_app'
