+++ website/pages/en/_meta.js
@@ -28,10 +28,16 @@ export default {
   },
   '###2': {
     type: 'heading',
+    title: 'Substreams',
+  },
+  substreams: {
+    type: 'children'
+  },
+  '###3': {
+    type: 'heading',
     title: 'Indexing',
   },
   'operating-graph-node': '',
-  substreams: 'Substreams',
   firehose: '',
   '---4': {
     type: 'separator',
