+++ website/pages/en/substreams/[[...slug]].mdx
@@ -2,7 +2,7 @@ import { visit } from 'unist-util-visit'
 import { RemoteContent } from 'nextra/data'
 import { buildDynamicMDX } from 'nextra/remote'
 import { listFiles } from './_meta.js'
-import { getPageMap } from '@/components/get-page-map'
+import { getPageMap } from '@/src/get-page-map'
 
 export async function getStaticPaths() {
   const files = await listFiles()
