+++ website/pages/ja/_meta.js
@@ -1,7 +1,7 @@
 import meta from '../en/_meta.js'
 
 export default {
-  ...meta,
+  ...JSON.parse(JSON.stringify(meta)),
   network: 'グラフネットワーク',
   '###1': {
     type: 'heading',
@@ -13,7 +13,7 @@ export default {
   managing: '管理します',
   querying: 'クエリ',
   cookbook: 'クックブック',
-  '###2': {
+  '###3': {
     type: 'heading',
     title: 'インデキシング',
   },
